added numbers to tables and figures
added the numbers at the end of table 9. Additionally, added solid lines for the respective mean values in figures 15 and 18.
spotted a mistake: In table 9 the columns precision and recall were switched.

The \orcid-command inserted below author email only produced a warning, changed it to ORCID: xxxx and added orcids